# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic SDLC Platform API
  description: REST API for the multi-agent SDLC orchestration platform.
  version: 0.4.0
...
paths:
...
  /api/milestones:
    post:
      tags:
        - milestones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - status
                - progress
                - completion
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                status:
                  type: string
                  enum:
                    - not started
                    - in progress
                    - complete
                    - blocked
                progress:
                  type: integer
                  minimum: 0
                  maximum: 100
                completion:
                  type: string
                  pattern: ^\d{4}-\d{2}-\d{2}$
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
    get:
      tags:
        - milestones
      responses:
        "200":
          description: Default Response
  /api/milestones/{id}:
    get:
      tags:
        - milestones
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    put:
      tags:
        - milestones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                status:
                  type: string
                  enum:
                    - not started
                    - in progress
                    - complete
                    - blocked
                progress:
                  type: integer
                  minimum: 0
                  maximum: 100
                completion:
                  type: string
                  pattern: ^\d{4}-\d{2}-\d{2}$
              additionalProperties: false
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                description: Not found
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
  /api/milestones/{id}/archive:
    patch:
      tags:
        - milestones
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "404":
          description: Not found
          content:
            application/json:
              schema:
                description: Not found
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
  /api/milestone-templates:
    post:
      tags:
        - milestones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - milestones
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                milestones:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - name
                      - status
                      - progress
                      - completion
                    properties:
                      name:
                        type: string
                        minLength: 1
                        maxLength: 255
                      status:
                        type: string
                        enum:
                          - not started
                          - in progress
                          - complete
                          - blocked
                      progress:
                        type: integer
                        minimum: 0
                        maximum: 100
                      completion:
                        type: string
                        pattern: ^\d{4}-\d{2}-\d{2}$
                    additionalProperties: false
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: Success
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                description: Validation error
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                description: Forbidden
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                description: Rate limited
    get:
      tags:
        - milestones
      responses:
        "200":
          description: Default Response
  /api/milestone-templates/{id}:
    delete:
      tags:
        - milestones
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /api/milestone-templates/{id}/apply:
    post:
      tags:
        - milestones
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: http://localhost:3000
    description: Local development server
...